requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
opencv-python==4.5.1.48
pafy==0.5.5
Pillow==8.1.2
pytesseract==0.3.7
requests==2.25.1
urllib3==1.26.3
Werkzeug==1.0.1
youtube-dl==2021.3.14
